refactor: Organize migration tests
