+++ crates/sui-genesis-builder/src/stardust/migration/migration.rs
@@ -1,10 +1,30 @@
+// Copyright (c) 2024 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
 //! Contains the logic for the migration process.
-use move_core_types::{ident_str, language_storage::StructTag};
 use std::{
     collections::{BTreeSet, HashMap},
     io::{BufWriter, Write},
     sync::Arc,
 };
+
+use anyhow::Result;
+use fastcrypto::hash::HashFunction;
+use iota_sdk::types::block::output::{
+    AliasOutput as StardustAlias, BasicOutput, FoundryOutput, NativeTokens, NftOutput, Output,
+    TokenId, TreasuryOutput,
+};
+use move_core_types::{ident_str, language_storage::StructTag};
+use move_vm_runtime_v2::move_vm::MoveVM;
+
+use sui_adapter_v2::{
+    adapter::new_move_vm, gas_charger::GasCharger, programmable_transactions,
+    temporary_store::TemporaryStore,
+};
+use sui_framework::BuiltInFramework;
+use sui_move_build::CompiledPackage;
+use sui_move_natives_v2::all_natives;
+use sui_protocol_config::{Chain, ProtocolConfig, ProtocolVersion};
 use sui_types::{
     balance::Balance,
     base_types::{ObjectRef, SequenceNumber},
@@ -15,22 +35,6 @@ use sui_types::{
     transaction::{Argument, InputObjects, ObjectArg},
     TypeTag, TIMELOCK_PACKAGE_ID,
 };
-
-use anyhow::Result;
-use fastcrypto::hash::HashFunction;
-use iota_sdk::types::block::output::{
-    AliasOutput as StardustAlias, BasicOutput, FoundryOutput, NativeTokens, NftOutput, Output,
-    TokenId, TreasuryOutput,
-};
-use move_vm_runtime_v2::move_vm::MoveVM;
-use sui_adapter_v2::{
-    adapter::new_move_vm, gas_charger::GasCharger, programmable_transactions,
-    temporary_store::TemporaryStore,
-};
-use sui_framework::BuiltInFramework;
-use sui_move_build::CompiledPackage;
-use sui_move_natives_v2::all_natives;
-use sui_protocol_config::{Chain, ProtocolConfig, ProtocolVersion};
 use sui_types::{
     base_types::{ObjectID, SuiAddress, TxContext},
     crypto::DefaultHash,
@@ -48,10 +52,11 @@ use sui_types::{
     MOVE_STDLIB_PACKAGE_ID, STARDUST_PACKAGE_ID, SUI_FRAMEWORK_PACKAGE_ID, SUI_SYSTEM_PACKAGE_ID,
 };
 
-use super::types::{snapshot::OutputHeader, stardust_to_sui_address_owner, Alias, AliasOutput};
 use crate::process_package;
 use crate::stardust::native_token::package_builder;
 use crate::stardust::native_token::package_data::NativeTokenPackageData;
+use crate::stardust::types::snapshot::OutputHeader;
+use crate::stardust::types::{stardust_to_sui_address_owner, Alias, AliasOutput};
 
 /// The dependencies of the generated packages for native tokens.
 pub const PACKAGE_DEPS: [ObjectID; 5] = [
@@ -65,7 +70,7 @@ pub const PACKAGE_DEPS: [ObjectID; 5] = [
 /// We fix the protocol version used in the migration.
 pub const MIGRATION_PROTOCOL_VERSION: u64 = 42;
 
-const NATIVE_TOKEN_BAG_KEY_TYPE: &str = "0x01::ascii::String";
+pub(crate) const NATIVE_TOKEN_BAG_KEY_TYPE: &str = "0x01::ascii::String";
 
 /// The orchestrator of the migration process.
 ///
@@ -186,7 +191,7 @@ fn generate_package(foundry: &FoundryOutput) -> Result<CompiledPackage> {
 
 /// On-chain data about the objects created while
 /// publishing foundry packages
-struct FoundryLedgerData {
+pub(crate) struct FoundryLedgerData {
     minted_coin_id: ObjectID,
     coin_type_origin: TypeOrigin,
     package_id: ObjectID,
@@ -211,7 +216,7 @@ impl FoundryLedgerData {
 /// Creates the objects that map to the stardust UTXO ledger.
 ///
 /// Internally uses an unmetered Move VM.
-struct Executor {
+pub(crate) struct Executor {
     protocol_config: ProtocolConfig,
     tx_context: TxContext,
     /// Stores all the migration objects.
@@ -230,7 +235,7 @@ struct Executor {
 impl Executor {
     /// Setup the execution environment backed by an in-memory store that holds
     /// all the system packages.
-    fn new(protocol_version: ProtocolVersion) -> Result<Self> {
+    pub(crate) fn new(protocol_version: ProtocolVersion) -> Result<Self> {
         let mut tx_context = create_migration_context();
         // Use a throwaway metrics registry for transaction execution.
         let metrics = Arc::new(LimitsMetrics::new(&prometheus::Registry::new()));
@@ -279,9 +284,24 @@ impl Executor {
         })
     }
 
+    #[cfg(test)]
+    pub(crate) fn store(&self) -> &InMemoryStorage {
+        &self.store
+    }
+
+    #[cfg(test)]
+    pub(crate) fn store_mut(&mut self) -> &mut InMemoryStorage {
+        &mut self.store
+    }
+
+    #[cfg(test)]
+    pub(crate) fn native_tokens(&mut self) -> &HashMap<TokenId, FoundryLedgerData> {
+        &self.native_tokens
+    }
+
     /// Load input objects from the store to be used as checked
     /// input while executing a transaction
-    fn load_input_objects(
+    pub(crate) fn load_input_objects(
         &self,
         object_refs: impl IntoIterator<Item = ObjectRef> + 'static,
     ) -> impl Iterator<Item = ObjectReadResult> + '_ {
@@ -295,7 +315,7 @@ impl Executor {
 
     /// Load packages from the store to be used as checked
     /// input while executing a transaction
-    fn load_packages(
+    pub(crate) fn load_packages(
         &self,
         object_ids: impl IntoIterator<Item = ObjectID> + 'static,
     ) -> impl Iterator<Item = ObjectReadResult> + '_ {
@@ -311,7 +331,7 @@ impl Executor {
         CheckedInputObjects::new_for_genesis(self.load_packages(PACKAGE_DEPS).collect())
     }
 
-    fn execute_pt_unmetered(
+    pub(crate) fn execute_pt_unmetered(
         &mut self,
         input_objects: CheckedInputObjects,
         pt: ProgrammableTransaction,
@@ -344,7 +364,7 @@ impl Executor {
     /// * For each native token, map the [`TokenId`] to the [`ObjectID`] of the
     ///   coin that holds its total supply.
     /// * Update the inner store with the created objects.
-    fn create_foundries(
+    pub(crate) fn create_foundries(
         &mut self,
         foundries: impl Iterator<Item = (FoundryOutput, CompiledPackage)>,
     ) -> Result<()> {
@@ -462,7 +482,7 @@ impl Executor {
     }
 
     /// Create a [`Bag`] of balances of native tokens executing a programmable transaction block.
-    fn create_bag_with_pt(
+    pub(crate) fn create_bag_with_pt(
         &mut self,
         native_tokens: &NativeTokens,
     ) -> Result<(Bag, SequenceNumber)> {
@@ -603,7 +623,8 @@ impl Executor {
         header: OutputHeader,
         basic_output: BasicOutput,
     ) -> Result<()> {
-        let mut data = super::types::output::BasicOutput::new(header.clone(), &basic_output);
+        let mut data =
+            crate::stardust::types::output::BasicOutput::new(header.clone(), &basic_output);
         let owner: SuiAddress = basic_output.address().to_string().parse()?;
 
         // The minimum version of the manually created objects
@@ -737,34 +758,7 @@ mod pt {
 
 #[cfg(test)]
 mod tests {
-    use std::str::FromStr;
-
-    use crate::stardust::{
-        migration::{Executor, Migration, MIGRATION_PROTOCOL_VERSION},
-        types::{snapshot::OutputHeader, Alias, AliasOutput, ALIAS_OUTPUT_MODULE_NAME},
-    };
-    use iota_sdk::types::block::{
-        address::AliasAddress,
-        address::{Address, Ed25519Address},
-        output::{
-            feature::{IssuerFeature, MetadataFeature, SenderFeature},
-            unlock_condition::{
-                GovernorAddressUnlockCondition, StateControllerAddressUnlockCondition,
-            },
-            AliasOutput as StardustAlias, AliasOutputBuilder, Feature,
-        },
-        output::{
-            unlock_condition::ImmutableAliasAddressUnlockCondition, AliasId, FoundryOutputBuilder,
-            NativeToken, SimpleTokenScheme, UnlockCondition,
-        },
-    };
     use sui_types::object::Object;
-    use sui_types::{
-        dynamic_field::{derive_dynamic_field_id, Field},
-        object::Owner,
-    };
-
-    use crate::stardust::native_token::package_data::NativeTokenModuleData;
 
     use super::*;
 
@@ -777,412 +771,5 @@ mod tests {
         create_snapshot(objects.clone(), &mut persisted).unwrap();
         let snapshot_objects: Vec<Object> = bcs::from_bytes(&persisted).unwrap();
         assert_eq!(objects, snapshot_objects);
-    }
-
-    fn random_output_header() -> OutputHeader {
-        OutputHeader::new_testing(
-            rand::random(),
-            rand::random(),
-            rand::random(),
-            rand::random(),
-        )
-    }
-
-    fn run_migration(outputs: impl IntoIterator<Item = (OutputHeader, Output)>) -> Vec<Object> {
-        let mut snapshot_buffer = Vec::new();
-        let mut foundries = Vec::new();
-        let mut outputs_without_foundries = Vec::new();
-
-        for (header, output) in outputs.into_iter() {
-            match output {
-                Output::Foundry(foundry) => {
-                    foundries.push((header, foundry));
-                }
-                other => {
-                    outputs_without_foundries.push((header, other));
-                }
-            }
-        }
-
-        Migration::new()
-            .unwrap()
-            .run(
-                foundries.into_iter(),
-                outputs_without_foundries.into_iter(),
-                &mut snapshot_buffer,
-            )
-            .unwrap();
-
-        bcs::from_bytes(&snapshot_buffer).unwrap()
-    }
-
-    fn migrate_alias(
-        header: OutputHeader,
-        stardust_alias: StardustAlias,
-    ) -> (ObjectID, Alias, AliasOutput) {
-        let alias_id: AliasId = stardust_alias
-            .alias_id()
-            .or_from_output_id(&header.output_id())
-            .to_owned();
-        let mut snapshot_buffer = Vec::new();
-        Migration::new()
-            .unwrap()
-            .run(
-                [].into_iter(),
-                [(header, stardust_alias.into())].into_iter(),
-                &mut snapshot_buffer,
-            )
-            .unwrap();
-
-        let migrated_objects: Vec<Object> = bcs::from_bytes(&snapshot_buffer).unwrap();
-
-        // Ensure the migrated objects exist under the expected identifiers.
-        let alias_object_id = ObjectID::new(*alias_id);
-        let alias_object = migrated_objects
-            .iter()
-            .find(|obj| obj.id() == alias_object_id)
-            .expect("alias object should be present in the migrated snapshot");
-        assert_eq!(alias_object.struct_tag().unwrap(), Alias::tag(),);
-        let alias_output_object = migrated_objects
-            .iter()
-            .find(|obj| match obj.struct_tag() {
-                Some(tag) => tag == AliasOutput::tag(),
-                None => false,
-            })
-            .expect("alias object should be present in the migrated snapshot");
-
-        // Version is set to 1 when the alias is created based on the computed lamport timestamp.
-        // When the alias is attached to the alias output, the version should be incremented.
-        assert!(
-            alias_object.version().value() > 1,
-            "alias object version should have been incremented"
-        );
-        assert!(
-            alias_output_object.version().value() > 1,
-            "alias output object version should have been incremented"
-        );
-
-        let alias_output: AliasOutput =
-            bcs::from_bytes(alias_output_object.data.try_as_move().unwrap().contents()).unwrap();
-        let alias: Alias =
-            bcs::from_bytes(alias_object.data.try_as_move().unwrap().contents()).unwrap();
-
-        (alias_object_id, alias, alias_output)
-    }
-
-    /// Test that the migrated alias objects in the snapshot contain the expected data.
-    #[test]
-    fn test_alias_migration() {
-        let alias_id = AliasId::new(rand::random());
-        let random_address = Ed25519Address::from(rand::random::<[u8; Ed25519Address::LENGTH]>());
-        let header = random_output_header();
-
-        let stardust_alias = AliasOutputBuilder::new_with_amount(1_000_000, alias_id)
-            .add_unlock_condition(StateControllerAddressUnlockCondition::new(random_address))
-            .add_unlock_condition(GovernorAddressUnlockCondition::new(random_address))
-            .with_state_metadata([0xff; 1])
-            .with_features(vec![
-                Feature::Metadata(MetadataFeature::new([0xdd; 1]).unwrap()),
-                Feature::Sender(SenderFeature::new(random_address)),
-            ])
-            .with_immutable_features(vec![
-                Feature::Metadata(MetadataFeature::new([0xaa; 1]).unwrap()),
-                Feature::Issuer(IssuerFeature::new(random_address)),
-            ])
-            .with_state_index(3)
-            .finish()
-            .unwrap();
-
-        let (alias_object_id, alias, alias_output) = migrate_alias(header, stardust_alias.clone());
-        let expected_alias = Alias::try_from_stardust(alias_object_id, &stardust_alias).unwrap();
-
-        // Compare only the balance. The ID is newly generated and the bag is tested separately.
-        assert_eq!(stardust_alias.amount(), alias_output.iota.value());
-
-        assert_eq!(expected_alias, alias);
-    }
-
-    /// Test that an Alias with a zeroed ID is migrated to an Alias Object with its UID set to the hashed Output ID.
-    #[test]
-    fn test_alias_migration_with_zeroed_id() {
-        let random_address = Ed25519Address::from(rand::random::<[u8; Ed25519Address::LENGTH]>());
-        let header = random_output_header();
-
-        let stardust_alias = AliasOutputBuilder::new_with_amount(1_000_000, AliasId::null())
-            .add_unlock_condition(StateControllerAddressUnlockCondition::new(random_address))
-            .add_unlock_condition(GovernorAddressUnlockCondition::new(random_address))
-            .finish()
-            .unwrap();
-
-        // If this function does not panic, then the created aliases
-        // were found at the correct non-zeroed Alias ID.
-        migrate_alias(header, stardust_alias);
-    }
-
-    /// Test that an Alias owned by another Alias can be received by the owning object.
-    ///
-    /// The PTB sends the extracted assets to the null address since it must be used in the transaction.
-    #[test]
-    fn test_alias_migration_with_alias_owner() {
-        let random_address = Ed25519Address::from(rand::random::<[u8; Ed25519Address::LENGTH]>());
-
-        let alias1_amount = 1_000_000;
-        let stardust_alias1 =
-            AliasOutputBuilder::new_with_amount(alias1_amount, AliasId::new(rand::random()))
-                .add_unlock_condition(StateControllerAddressUnlockCondition::new(random_address))
-                .add_unlock_condition(GovernorAddressUnlockCondition::new(random_address))
-                .finish()
-                .unwrap();
-
-        let alias2_amount = 2_000_000;
-        // stardust_alias1 is the owner of stardust_alias2.
-        let stardust_alias2 =
-            AliasOutputBuilder::new_with_amount(alias2_amount, AliasId::new(rand::random()))
-                .add_unlock_condition(StateControllerAddressUnlockCondition::new(Address::from(
-                    stardust_alias1.alias_id().clone(),
-                )))
-                .add_unlock_condition(GovernorAddressUnlockCondition::new(Address::from(
-                    stardust_alias1.alias_id().clone(),
-                )))
-                .finish()
-                .unwrap();
-
-        let migrated_objects = run_migration([
-            (random_output_header(), stardust_alias1.into()),
-            (random_output_header(), stardust_alias2.into()),
-        ]);
-
-        // Find the corresponding objects to the migrated aliases, uniquely identified by their amounts.
-        // Should be adapted to use the tags from issue 239 to make this much easier.
-        let alias_output1_id = migrated_objects
-            .iter()
-            .find(|obj| {
-                obj.struct_tag()
-                    .map(|tag| tag == AliasOutput::tag())
-                    .unwrap_or(false)
-                    && bcs::from_bytes::<AliasOutput>(obj.data.try_as_move().unwrap().contents())
-                        .unwrap()
-                        .iota
-                        .value()
-                        == alias1_amount
-            })
-            .expect("alias1 should exist")
-            .id();
-
-        let alias_output2_id = migrated_objects
-            .iter()
-            .find(|obj| {
-                obj.struct_tag()
-                    .map(|tag| tag == AliasOutput::tag())
-                    .unwrap_or(false)
-                    && bcs::from_bytes::<AliasOutput>(obj.data.try_as_move().unwrap().contents())
-                        .unwrap()
-                        .iota
-                        .value()
-                        == alias2_amount
-            })
-            .expect("alias2 should exist")
-            .id();
-
-        let mut executor = Executor::new(MIGRATION_PROTOCOL_VERSION.into()).unwrap();
-        for object in migrated_objects {
-            executor.store.insert_object(object);
-        }
-
-        let alias_output1_object_ref = executor
-            .store
-            .get_object(&alias_output1_id)
-            .unwrap()
-            .compute_object_reference();
-
-        let alias_output2_object_ref = executor
-            .store
-            .get_object(&alias_output2_id)
-            .unwrap()
-            .compute_object_reference();
-
-        let pt = {
-            let mut builder = ProgrammableTransactionBuilder::new();
-            let alias1_arg = builder
-                .obj(ObjectArg::ImmOrOwnedObject(alias_output1_object_ref))
-                .unwrap();
-
-            let extracted_assets = builder.programmable_move_call(
-                STARDUST_PACKAGE_ID,
-                ALIAS_OUTPUT_MODULE_NAME.into(),
-                ident_str!("extract_assets").into(),
-                vec![],
-                vec![alias1_arg],
-            );
-
-            let Argument::Result(result_idx) = extracted_assets else {
-                panic!("expected Argument::Result");
-            };
-            let balance_arg = Argument::NestedResult(result_idx, 0);
-            let bag_arg = Argument::NestedResult(result_idx, 1);
-            let alias1_arg = Argument::NestedResult(result_idx, 2);
-
-            let receiving_alias2_arg = builder
-                .obj(ObjectArg::Receiving(alias_output2_object_ref))
-                .unwrap();
-            let received_alias_output2 = builder.programmable_move_call(
-                STARDUST_PACKAGE_ID,
-                ident_str!("address_unlock_condition").into(),
-                ident_str!("unlock_alias_address_owned_alias").into(),
-                vec![],
-                vec![alias1_arg, receiving_alias2_arg],
-            );
-
-            let coin_arg = builder.programmable_move_call(
-                SUI_FRAMEWORK_PACKAGE_ID,
-                ident_str!("coin").into(),
-                ident_str!("from_balance").into(),
-                vec![
-                    StructTag::from_str(&format!("{}::sui::SUI", SUI_FRAMEWORK_PACKAGE_ID))
-                        .unwrap()
-                        .into(),
-                ],
-                vec![balance_arg],
-            );
-
-            builder.transfer_arg(SuiAddress::default(), bag_arg);
-            builder.transfer_arg(SuiAddress::default(), coin_arg);
-
-            // We have to use Alias Output as we cannot transfer it (since it lacks the `store` ability),
-            // so we extract its assets.
-            let extracted_assets = builder.programmable_move_call(
-                STARDUST_PACKAGE_ID,
-                ALIAS_OUTPUT_MODULE_NAME.into(),
-                ident_str!("extract_assets").into(),
-                vec![],
-                vec![received_alias_output2],
-            );
-            let Argument::Result(result_idx) = extracted_assets else {
-                panic!("expected Argument::Result");
-            };
-            let balance_arg = Argument::NestedResult(result_idx, 0);
-            let bag_arg = Argument::NestedResult(result_idx, 1);
-            let alias2_arg = Argument::NestedResult(result_idx, 2);
-
-            let coin_arg = builder.programmable_move_call(
-                SUI_FRAMEWORK_PACKAGE_ID,
-                ident_str!("coin").into(),
-                ident_str!("from_balance").into(),
-                vec![
-                    StructTag::from_str(&format!("{}::sui::SUI", SUI_FRAMEWORK_PACKAGE_ID))
-                        .unwrap()
-                        .into(),
-                ],
-                vec![balance_arg],
-            );
-
-            builder.transfer_arg(SuiAddress::default(), coin_arg);
-            builder.transfer_arg(SuiAddress::default(), bag_arg);
-
-            builder.transfer_arg(SuiAddress::default(), alias1_arg);
-            builder.transfer_arg(SuiAddress::default(), alias2_arg);
-
-            builder.finish()
-        };
-
-        let input_objects = CheckedInputObjects::new_for_genesis(
-            executor
-                .load_input_objects([alias_output1_object_ref])
-                .chain(executor.load_packages(PACKAGE_DEPS))
-                .collect(),
-        );
-        executor.execute_pt_unmetered(input_objects, pt).unwrap();
-    }
-
-    #[test]
-    fn create_bag_with_pt() {
-        // Mock the foundry
-        let owner = AliasAddress::new(AliasId::new([0; AliasId::LENGTH]));
-        let supply = 1_000_000;
-        let token_scheme = SimpleTokenScheme::new(supply, 0, supply).unwrap();
-        let foundry = FoundryOutputBuilder::new_with_amount(1000, 1, token_scheme.into())
-            .with_unlock_conditions([UnlockCondition::from(
-                ImmutableAliasAddressUnlockCondition::new(owner),
-            )])
-            .finish_with_params(supply)
-            .unwrap();
-        let foundry_id = foundry.id();
-        let foundry_package_data = NativeTokenPackageData::new(
-            "wat",
-            NativeTokenModuleData::new(
-                foundry_id, "wat", "WAT", 0, "WAT", supply, supply, "wat", "wat", None, owner,
-            ),
-        );
-        let foundry_package = package_builder::build_and_compile(foundry_package_data).unwrap();
-
-        // Execution
-        let mut executor = Executor::new(ProtocolVersion::MAX).unwrap();
-        let object_count = executor.store.objects().len();
-        executor
-            .create_foundries([(foundry, foundry_package)].into_iter())
-            .unwrap();
-        // Foundry package publication creates four objects
-        //
-        // * The package
-        // * Coin metadata
-        // * MaxSupplyPolicy
-        // * The total supply coin
-        assert_eq!(executor.store.objects().len() - object_count, 4);
-        assert!(executor.native_tokens.get(&foundry_id.into()).is_some());
-        let initial_supply_coin_object = executor
-            .store
-            .objects()
-            .values()
-            .find_map(|object| object.is_coin().then_some(object))
-            .expect("there should be only a single coin: the total supply of native tokens");
-        let coin_type_tag = initial_supply_coin_object.coin_type_maybe().unwrap();
-        let initial_supply_coin_data = initial_supply_coin_object.as_coin_maybe().unwrap();
-
-        // Mock the native token
-        let token_amount = 10_000;
-        let native_token = NativeToken::new(foundry_id.into(), token_amount).unwrap();
-
-        // Create the bag
-        let (bag, _) = executor
-            .create_bag_with_pt(&NativeTokens::from_vec(vec![native_token]).unwrap())
-            .unwrap();
-        assert!(executor.store.get_object(bag.id.object_id()).is_none());
-
-        // Verify the mutation of the foundry coin with the total supply
-        let mutated_supply_coin = executor
-            .store
-            .get_object(initial_supply_coin_data.id())
-            .unwrap()
-            .as_coin_maybe()
-            .unwrap();
-        assert_eq!(mutated_supply_coin.value(), supply - token_amount);
-
-        // Get the dynamic fields (df)
-        let tokens = executor
-            .store
-            .objects()
-            .values()
-            .filter_map(|object| object.is_child_object().then_some(object))
-            .collect::<Vec<_>>();
-        assert_eq!(tokens.len(), 1);
-        assert_eq!(
-            tokens[0].owner,
-            Owner::ObjectOwner((*bag.id.object_id()).into())
-        );
-        let token_as_df = tokens[0].to_rust::<Field<String, Balance>>().unwrap();
-        // Verify name
-        let expected_name = coin_type_tag.to_canonical_string(true);
-        assert_eq!(token_as_df.name, expected_name);
-        // Verify value
-        let expected_balance = Balance::new(token_amount);
-        assert_eq!(token_as_df.value, expected_balance);
-        // Verify df id
-        let expected_id = derive_dynamic_field_id(
-            *bag.id.object_id(),
-            &NATIVE_TOKEN_BAG_KEY_TYPE.parse().unwrap(),
-            &bcs::to_bytes(&expected_name).unwrap(),
-        )
-        .unwrap();
-        assert_eq!(*token_as_df.id.object_id(), expected_id);
     }
 }
