+++ crates/sui-genesis-builder/src/stardust/migration/mod.rs
@@ -1,0 +1,6 @@
+// Copyright (c) 2024 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+pub mod migration;
+#[cfg(test)]
+mod tests;
